pyproject.toml
[project]
name = "ai-chat-frontend"
version = "0.1.0"
description = "Frontend service for AI Chat Platform"
authors = [
    {name = "Your Name", email = "you@example.com"},
]
dependencies = [
    "streamlit>=1.31.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.6.0",
    "streamlit-oauth>=0.1.0",
    "jwt>=1.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "ruff>=0.1.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src"

[tool.black]
line-length = 120
target-version = ["py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 120

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 120
target-version = "py310"